pyproject.toml
[project]
name = "base"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.123.5",
    "langchain>=1.1.0",
    "langchain-google-genai>=3.2.0",
    "langgraph>=1.0.4",
    "motor>=3.7.1",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
    "reportlab>=4.2.0",
]
